Added acorn.playInFrame, updated test files.
acorn.playInFrame plays given acornModel in given iframe.
If the iframe has no acorn object, it attaches an onload
callback to play as soon as the iframe finishes loading.

Warning: if the iframe has loaded but has no acorn object,
the onload will never fire and this call will be a noop.