rangeSlider: stop handles at element boundaries
NOTE: when initializing rangeslider widget on an element, that element
must either be in the DOM or get added to the DOM during that call
stack in order to render completely on its own. If the element is added
to the DOM at some later point (and not reinitialized), a hack will be
necessary to render slider values. Setting these values anew will
suffice.