ContentView: changed to use combined model
The main Player views (PlayerView, EditView, ContentView), all need
access to both the acornModel and the shellModel. Rather than setting
one of them as `@model`, and the other as `@nameModel`, I think it
makes more sense to pass them in through a combined object. Like so:

    new PlayerView
      model:
        acornModel: acornModel
        shellModel: shellModel

So, within the views, one can access each like so:

    @model.acornModel
    @model.shellModel

(Note that Backbone.View models need not be Backbone.Model instances)