models: clone deep-copies
Underscore's _.clone does not deep-copy.
(See https://github.com/documentcloud/underscore/issues/162 as to why).
This is a problem for our objects because they are objects with many
levels of depth. And for editing, it's really important the data of the
edit staging clone does not inadvertently modify the real object.