added .first() all over the place.
Using `@$('.classname')` is dangerous. if view is included recursively,
Bad Things (tm) will happen. The state is getting messed with in the
entire DOM subtree.

It seems that in most cases, adding `.first()` to the end of the
selector should fix it. (Is there a better approach than this!? it
seems awful...)

@tenedor could you verify that these don't BreakEverything? I looked
at each one and these seemed fine. I left some other ones that were
dubious (i.e. multiple divs are wanted in the selector).

@ali01 fyi.